Våra fyrbenta vänner

+++ src/Husdjurshotellet/Djur.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 abstract public class Djur implements ICalcFood {
     protected String namn;
-    protected double vikt;
+    protected int vikt;
     DjurFoder matVanor;
     static ArrayList<Djur> djurlista = new ArrayList<Djur>();
